[platform_device_loader] isolate async_enrollment related operations
